requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.1
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
loguru==0.7.0
pycparser==2.21
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
